+++ src/main/java/com/drmangotea/tfmg/content/machinery/oil_processing/pumpjack/base/PumpjackBaseBlockEntity.java
@@ -134,7 +134,7 @@ public class PumpjackBaseBlockEntity extends SmartBlockEntity implements IHaveGo
 
         if (tank.getFluidAmount() + miningRate > tank.getCapacity())
             return;
-        if (!TFMGConfigs.common().worldgen.infiniteDeposits.get() && reservoir.isEmpty()) {
+        if (!TFMGConfigs.common().worldGen.infiniteDeposits.get() && reservoir.isEmpty()) {
             deposit = null;
             findDeposit();
             return;
@@ -144,7 +144,7 @@ public class PumpjackBaseBlockEntity extends SmartBlockEntity implements IHaveGo
         if (amountPumped == 0)
             return;
 
-        if (TFMGConfigs.common().worldgen.infiniteDeposits.get())
+        if (TFMGConfigs.common().worldGen.infiniteDeposits.get())
             return;
 
         RandomSource randomSource = level.getRandom();